(PC-35270) remarks after review
