+++ src/features/offer/fixtures/offerImageDimensions.ts
@@ -1,4 +1,4 @@
-import { OfferImageContainerDimensions } from 'features/offer/helpers/useOfferImageContainerDimensions'
+import { OfferImageContainerDimensions } from 'features/offer/types'
 
 export const mockOfferImageDimensions: OfferImageContainerDimensions = {
   backgroundHeight: 100,
@@ -8,12 +8,5 @@ export const mockOfferImageDimensions: OfferImageContainerDimensions = {
     maxWidth: 300,
     aspectRatio: 2 / 3,
     borderRadius: 8,
-  },
-  imageStyleWithoutBorderRadius: {
-    height: 200,
-    width: 150,
-    maxWidth: 300,
-    aspectRatio: 2 / 3,
-    borderRadius: 0,
   },
 }
